+++ client/.prettierrc.json
@@ -1,4 +1,3 @@
-
 {
   "$schema": "https://json.schemastore.org/prettierrc",
   "semi": false,
