+++ client/.vscode/extensions.json
