+++ client/README.md
@@ -1,0 +1,51 @@
+# Frontend
+
+Soon...
+
+## Docs
+
+### UI Kit
+
+We are using [PrimeVue](https://primevue.org/introduction/) as main UI-kit, and default `<style>` tag for create needed styles
+
+### Recommended IDE Setup
+
+[VSCode](https://code.visualstudio.com/) + [Volar](https://marketplace.visualstudio.com/items?itemName=Vue.volar) (and disable Vetur).
+
+### Type Support for `.vue` Imports in TS
+
+TypeScript cannot handle type information for `.vue` imports by default, so we replace the `tsc` CLI with `vue-tsc` for type checking. In editors, we need [Volar](https://marketplace.visualstudio.com/items?itemName=Vue.volar) to make the TypeScript language service aware of `.vue` types.
+
+### Customize configuration
+
+See [Vite Configuration Reference](https://vite.dev/config/).
+
+### Project Setup
+
+```sh
+npm install
+```
+
+#### Compile and Hot-Reload for Development
+
+```sh
+npm run dev
+```
+
+#### Type-Check, Compile and Minify for Production
+
+```sh
+npm run build
+```
+
+#### Lint with [ESLint](https://eslint.org/)
+
+```sh
+npm run lint
+```
+
+#### Format your code with Prettier config
+
+```sh
+npm run format
+```
