+++ client/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "frontend",
-  "version": "0.1.2",
+  "name": "client",
+  "version": "0.2.0",
   "private": true,
   "type": "module",
   "scripts": {
@@ -10,9 +10,11 @@
     "build-only": "vite build",
     "type-check": "vue-tsc --build --force",
     "lint": "eslint . --fix",
-    "format": "prettier --write src/"
+    "format": "prettier --write ."
   },
   "dependencies": {
+    "@primevue/themes": "^4.2.5",
+    "primevue": "^4.2.5",
     "vue": "^3.5.12",
     "vue-router": "^4.4.5"
   },
