+++ client/src/App.vue
@@ -17,13 +17,7 @@ export default {
 <style>
 * {
   box-sizing: border-box;
-}
-
-body {
   margin: 0;
   padding: 0;
-  font-family: Roboto, sans-serif;
-
-  background-color: white;
 }
 </style>
