+++ client/src/main.ts
@@ -1,0 +1,22 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+
+import PrimeVue from 'primevue/config'
+import Aura from '@primevue/themes/aura'
+
+import './main.css'
+
+const app = createApp(App)
+
+app.use(PrimeVue, {
+  theme: {
+    preset: Aura,
+    options: {
+      prefix: 'p',
+      darkModeSelector: 'system',
+      cssLayer: false,
+    },
+  },
+})
+
+app.mount('#app')
