+++ client/src/types/index.ts
@@ -1,0 +1,4 @@
+export type Item = {
+  title: string
+  description: string
+}
